Support variable tile dimensions

+++ include/noworry/Draw2D/SpriteRenderer.h
@@ -14,10 +14,10 @@
 
 struct Region
 {
-    int x;
-    int y;
-    int w;
-    int h;
+    float x;
+    float y;
+    float w;
+    float h;
 };
 
 // TODO: Investigate more efficient sprite renderer implementation
